Readme updated, new dummy test file added, refactoring
